requirements.txt
# === Main ===
fastapi==0.115.0
uvicorn[standard]==0.30.1
pydantic==2.8.2
pydantic-settings==2.3.4
sqlalchemy[asyncio]==2.0.36
asyncpg==0.29.0
alembic==1.13.2
psycopg[binary]==3.2.1
redis==5.0.7
python-dotenv==1.0.1
httpx==0.27.0
loguru==0.7.2
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
email-validator==2.2.0
python-multipart


# === Background tasks ===
celery[redis]==5.4.0

# === Observability ===
opentelemetry-api==1.28.2
opentelemetry-sdk==1.28.2
opentelemetry-instrumentation-fastapi==0.49b2
opentelemetry-exporter-otlp==1.28.2
prometheus-client==0.20.0

# === Security ===
itsdangerous==2.2.0
pyjwt[crypto]==2.9.0

# === Dev only ===
ruff==0.5.7
mypy==1.11.1
types-python-jose==3.3.4.8
pre-commit==3.7.1

# === Testing ===
pytest==8.3.2
pytest-asyncio==0.23.8
pytest-cov==5.0.0
factory-boy==3.3.0
faker==27.0.0